refactor(web): 评估面板 / SeriesChart 从 MonitorDashboard 抽出
纯搬家，零行为变化。评估的对象是 version 下的一组 checkpoint，不是某一次训练
进程，所以结果面板得能脱离训练任务页单独存在 —— 先把它从 MonitorDashboard 里
挪出来，后面的 commit 才好在版本页复用同一个组件而不是抄第二份。

SeriesChart（含 ChartSvg / calcEMA / downsample）两边都要用，一并抽成独立文件，
避免 EvalMetricsPanel 反过来 import MonitorDashboard 形成环。
