+++ studio/web/src/pages/QueueDetail.tsx
@@ -14,7 +14,8 @@ import { useDialog } from '../components/Dialog'
 import { useToast } from '../components/Toast'
 import { useEventStream } from '../lib/useEventStream'
 import { useTaskEvalProgress } from '../lib/useEvalProgress'
-import MonitorDashboard, { EvalMetricsPanel } from '../components/MonitorDashboard'
+import MonitorDashboard from '../components/MonitorDashboard'
+import { EvalMetricsPanel } from '../components/EvalMetricsPanel'
 import TaskLogDrawer, { type LogSource, type LogSourceStatus } from '../components/TaskLogDrawer'
 import { useMonitorProgress } from '../lib/useMonitorProgress'
 import { taskKind } from './Queue'
